+++ src/index.ts
@@ -10,4 +10,5 @@ export import StartPoint = StartPoint.StartPoint;
 export import Permission = Permission.Permission;
 
 export import AMFlowError = Error.AMFlowError;
-export { promisifyAMFlow, PromisifiedAMFlow } from "./auxiliary/PromisifyAMFlow";
+export { PromisifiedAMFlow } from "./auxiliary/PromisifiedAMFlow";
+export { PromisifiedAMFlowProxy } from "./auxiliary/PromisifiedAMFlowProxy";